+++ vars/scmServiceRegistration.groovy
@@ -4,12 +4,12 @@ import groovy.json.JsonBuilder
 
 import groovy.json.JsonSlurper
 import java.util.concurrent.TimeUnit
-import java.util.regex.Pattern
 import okhttp3.MediaType
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody
 import okhttp3.Response
+import uk.gov.hmcts.contino.IPV4Validator
 /*--------------------------------------------------------------
 Groovy script to update the scm service consul record. It will
 crawl the infra to workout which webapps are currently deployed
@@ -26,19 +26,6 @@ class ConsulRecord {
   List Tags
   String Address
   Integer Port
-}
-
-class IPV4Validator {
-  private static final String IPV4_REGEX = /^((0|1\d?\d?|2[0-4]?\d?|25[0-5]?|[3-9]\d?)\.){3}(0|1\d?\d?|2[0-4]?\d?|25[0-5]?|[3-9]\d?)$/
-  private static final Pattern IPV4_PATTERN = Pattern.compile(IPV4_REGEX)
-
-  static boolean validate(String ipAddress) {
-    if (ipAddress == null || ipAddress.isEmpty()) {
-      return false
-    }
-
-    return IPV4_PATTERN.matcher(ipAddress).matches()
-  }
 }
 
 def call(environment) {
